Remove write queries to the DB
Simply inserting or removing organization members directly via SQL
queries is dangerous, as it affects the DB consistency. For example,
simply inserting a user into an organization does not update the
corresponding member count, which will then be displayed incorrectly on
the web page.

It is probably better to wait for the Gogs API to include the necessary
features.